feat(course): Course student classroom

+++ app/Http/Controllers/CourseStudent/CourseStudentController.php
@@ -18,6 +18,7 @@ use \App\Models\Course;
 use \App\Models\Teacher;
 use \App\Models\Classroom;
 use \App\Models\CourseStudent;
+use \App\Models\ClassRoomStudent;
 
 class CourseStudentController extends Controller
 {
@@ -30,10 +31,13 @@ class CourseStudentController extends Controller
     {
         $course = Course::where('uuid', $uuid)->first();
         $students =  Teacher::find(Auth::user()->id)->students;
+
+
+
         return Inertia::render('CourseStudent', [
             'course'     => $course,
             'students'   => $students,
-            "classrooms" => Classroom::where("course_id", $course->id)->get()
+            "classrooms" => Classroom::with("students")->where("course_id", $course->id)->get(),
             ]);
     }
 
@@ -42,13 +46,29 @@ class CourseStudentController extends Controller
         //course student
         $course  = Course::where('uuid', $request->input("course"))->first();
         $student = Student::where('uuid', $request->input("student"))->first();
+        $classroom = Classroom::where('uuid', $request->input("classroom"))->first();
 
-        $course->students()->detach($student->id);
-        $course->students()->attach($student->id);
+        if($student->courses->count() > 0){
+            $course->students()->detach($student->id);
+        }else{
+            $course->students()->attach($student->id);
+        }
 
+        if($student->classrooms->count() > 0){
+            $classroom->students()->detach($student->id);
+        }else{
+            $classroom->students()->attach($student->id);
+        }
+
+        $request->session()->flash('message', 'Saved successfully!');
 
 
         return Redirect::route('teacher-course-student', $course->uuid);
+    }
+
+    public function book(Request $request)
+    {
+        return Classroom::with("students")->where("id", $request->input("classroom"))->count();
     }
